Move logging and history-recording into transport.
